adicionou autenticação de usuário com passport-jwt e adicionou tratamento de erros relacionados as autorizações do usuário

+++ src/controllers/categoriesController.ts
@@ -37,6 +37,7 @@ export async function getCategoryByName(req:Request, res:Response){
 }
 
 export async function postCategory(req:Request, res:Response){
+    if(!req.user?.isAuthor) return res.status(403).json({message: 'Client lacks access rights to this content'});
     const categoryName = req.body.category;
 
     const category = await prisma.categories.create({
@@ -50,6 +51,7 @@ export async function postCategory(req:Request, res:Response){
 }
 
 export async function putCategory(req:Request, res:Response){
+    if(!req.user?.isAuthor) return res.status(403).json({message: 'Client lacks access rights to this content'});
     const categoryName = req.body.category;
     const id = `${req.params.categoryId}`;
 
@@ -62,6 +64,7 @@ export async function putCategory(req:Request, res:Response){
 }
 
 export async function deleteCategory(req:Request, res:Response){
+    if(!req.user?.isAuthor) return res.status(403).json({message: 'Client lacks access rights to this content'});
     const id = `${req.params.categoryId}`;
     const category = await prisma.categories.delete({
         where: {id: parseInt(id)}
