+++ src/controllers/postsController.ts
@@ -37,6 +37,7 @@ export async function getPostsByCategory(req:Request, res:Response){
 }
 
 export async function postPost(req:Request, res:Response){
+    if(!req.user?.isAuthor) return res.status(403).json({message: 'Client lacks access rights to this content'});
     const {title, text, categoryId} = req.body;
     const post = await prisma.posts.create({
         data: {
@@ -51,6 +52,7 @@ export async function postPost(req:Request, res:Response){
 }
 
 export async function putPost(req:Request, res:Response){
+    if(!req.user?.isAuthor) return res.status(403).json({message: 'Client lacks access rights to this content'});
     const id = `${req.params.postId}`;
     const {title, text} = req.body;
     const post = await prisma.posts.update({
@@ -66,6 +68,7 @@ export async function putPost(req:Request, res:Response){
 }
 
 export async function deletePost(req:Request, res:Response){
+    if(!req.user?.isAuthor) return res.status(403).json({message: 'Client lacks access rights to this content'});
     const id = `${req.params.postId}`;
     const post = await prisma.posts.delete({
         where: {id: parseInt(id)}
